move functions away from api.py to genome_functions, since they will be used in multiple scripts
